Allow parsing of HTML 5 documents with embedded Starkups files
If the document starts with an HTML 5 doctype and an element with the
CSS class `.skp-file`, then the text of the first such element will be
used as the Starkups document.  Otherwise, then the original input will
be used unmodified.  The intended purpose is to allow HTML files
exported by the viewer to be reopened in the viewer without requiring
the original file to exist separately.